+++ Sources/carton/main.swift
@@ -30,193 +30,6 @@
 //
 // This executable should be eventually removed once SwiftPM provides a way to express those requirements.
 
-import CartonHelpers
-import Foundation
-import SwiftToolchain
-
-struct CartonCommandError: Error & CustomStringConvertible {
-  init(_ description: String) {
-    self.description = description
-  }
-  var description: String
-}
-
-extension Foundation.Process {
-  internal static func checkRun(
-    _ executableURL: URL, arguments: [String], forwardExit: Bool = false
-  ) throws {
-    fputs(
-      "Running \(([executableURL.path] + arguments).map { "\"\($0)\"" }.joined(separator: " "))\n",
-      stderr)
-    fflush(stderr)
-
-    let process = Foundation.Process()
-    process.executableURL = executableURL
-    process.arguments = arguments
-
-    // Monitor termination/interrruption signals to forward them to child process
-    func setSignalForwarding(_ signalNo: Int32) {
-      signal(signalNo, SIG_IGN)
-      let signalSource = DispatchSource.makeSignalSource(signal: signalNo)
-      signalSource.setEventHandler {
-        signalSource.cancel()
-        process.interrupt()
-      }
-      signalSource.resume()
-    }
-    setSignalForwarding(SIGINT)
-    setSignalForwarding(SIGTERM)
-
-    if forwardExit {
-      process.terminationHandler = {
-        // Exit plugin process itself when child process exited
-        exit($0.terminationStatus)
-      }
-    }
-    try process.run()
-    process.waitUntilExit()
-
-    if process.terminationStatus != 0 {
-      exit(process.terminationStatus)
-    }
-  }
-}
-
-func derivePackageCommandArguments(
-  swiftExec: URL,
-  subcommand: String,
-  scratchPath: String,
-  extraArguments: [String]
-) throws -> [String] {
-  var packageArguments: [String] = [
-    "package", "--triple", "wasm32-unknown-wasi", "--scratch-path", scratchPath,
-  ]
-  let pluginArguments: [String] = ["plugin"]
-  var cartonPluginArguments: [String] = extraArguments
-
-  switch subcommand {
-  case "bundle":
-    packageArguments += ["--disable-sandbox"]
-    // TODO: Uncomment this line once we stop creating .carton directory in the home directory
-    // pluginArguments += ["--allow-writing-to-package-directory"]
-
-    // Place before user-given extra arguments to allow overriding default options
-    cartonPluginArguments = ["--output", "Bundle"] + cartonPluginArguments
-  case "dev":
-    packageArguments += ["--disable-sandbox"]
-  case "test":
-    // 1. Ask the plugin process to generate the build command based on the given options
-    let commandFile = try makeTemporaryFile(prefix: "test-build")
-    try Foundation.Process.checkRun(
-      swiftExec,
-      arguments: packageArguments + pluginArguments + [
-        "carton-test",
-        "internal-get-build-command",
-      ] + cartonPluginArguments + ["--output", commandFile.path]
-    )
-
-    // 2. Build the test product
-    let buildArguments = try String(contentsOf: commandFile).split(separator: "\n")
-    if !buildArguments.isEmpty {
-      let buildCommand = buildArguments.map(String.init) + [
-        // NOTE: "swift-build" uses llbuild manifest cache by default even though
-        // target triple changed.
-        "--disable-build-manifest-caching",
-        "--triple", "wasm32-unknown-wasi", "--scratch-path", scratchPath,
-      ]
-      try Foundation.Process.checkRun(
-        swiftExec,
-        arguments: buildCommand
-      )
-    }
-
-    // "--environment browser" launches a http server
-    packageArguments += ["--disable-sandbox"]
-  default: break
-  }
-
-  return packageArguments + pluginArguments + ["carton-\(subcommand)"] + cartonPluginArguments
-}
-
-var errnoString: String {
-  String(cString: strerror(errno))
-}
-
-var temporaryDirectory: URL {
-  URL(fileURLWithPath: NSTemporaryDirectory())
-}
-
-func makeTemporaryFile(prefix: String, in directory: URL? = nil) throws -> URL {
-  let directory = directory ?? temporaryDirectory
-  var template = directory.appendingPathComponent("\(prefix)XXXXXX").path
-  let result = try template.withUTF8 { template in
-    let copy = UnsafeMutableBufferPointer<CChar>.allocate(capacity: template.count + 1)
-    defer { copy.deallocate() }
-    template.copyBytes(to: copy)
-    copy[template.count] = 0
-    guard mkstemp(copy.baseAddress!) != -1 else {
-      let error = errnoString
-      throw CartonCommandError("Failed to make a temporary file at \(template): \(error)")
-    }
-    return String(cString: copy.baseAddress!)
-  }
-  return URL(fileURLWithPath: result)
-}
-
-func pluginSubcommand(subcommand: String, argv0: String, arguments: [String]) async throws {
-  let scratchPath = URL(fileURLWithPath: ".build/carton")
-  if FileManager.default.fileExists(atPath: scratchPath.path) {
-    try FileManager.default.createDirectory(at: scratchPath, withIntermediateDirectories: true)
-  }
-
-  let terminal = InteractiveWriter.stdout
-  let toolchainSystem = try ToolchainSystem(fileSystem: localFileSystem)
-  let (swiftPath, _) = try await toolchainSystem.inferSwiftPath(terminal)
-  let extraArguments = arguments
-
-  let swiftExec = URL(fileURLWithPath: swiftPath.pathString)
-  let pluginArguments = try derivePackageCommandArguments(
-    swiftExec: swiftExec,
-    subcommand: subcommand,
-    scratchPath: scratchPath.path,
-    extraArguments: extraArguments
-  )
-
-  try Foundation.Process.checkRun(swiftExec, arguments: pluginArguments, forwardExit: true)
-}
-
-func main(arguments: [String]) async throws {
-  let argv0 = arguments[0]
-  let arguments = arguments.dropFirst()
-  let pluginSubcommands = ["bundle", "dev", "test"]
-  let subcommands = pluginSubcommands + ["package", "--version"]
-  guard let subcommand = arguments.first, subcommands.contains(subcommand) else {
-    if arguments.first == "init" {
-      print(
-        "Warning: 'init' subcommand has been removed, use 'swift package init' and add 'carton' as a dependency in Package.swift instead."
-      )
-    }
-    print("Usage: swift run carton <subcommand> [options]")
-    print("Available subcommands: \(subcommands.joined(separator: ", "))")
-    exit(1)
-  }
-
-  switch subcommand {
-  case _ where pluginSubcommands.contains(subcommand):
-    try await pluginSubcommand(
-      subcommand: subcommand, argv0: argv0, arguments: Array(arguments.dropFirst()))
-  case "package":
-    let terminal = InteractiveWriter.stdout
-    let toolchainSystem = try ToolchainSystem(fileSystem: localFileSystem)
-    let (swiftPath, _) = try await toolchainSystem.inferSwiftPath(terminal)
-    try Foundation.Process.checkRun(
-      URL(fileURLWithPath: swiftPath.pathString),
-      arguments: ["package"] + arguments.dropFirst(), forwardExit: true
-    )
-  case "--version":
-    print(cartonVersion)
-  default: fatalError("Unimplemented subcommand!?")
-  }
-}
+import CartonModule
 
 try await main(arguments: CommandLine.arguments)
